# backend/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: NyumbaLink API
  version: "2.0.0"
  description: |
    ## NyumbaLink — AI-Powered Accommodation Brokerage System
    **Dar es Salaam, Tanzania**

    NyumbaLink connects **Seekers** (tenants) with verified **Landlords** across
    Dar es Salaam's five municipalities — Kinondoni, Ilala, Ubungo, Temeke, and Kigamboni.

    ### Authentication
    Protected endpoints require a **Bearer JWT** in the `Authorization` header.
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `POST /api/v1/auth/login`.
    Access tokens expire after **24 hours**.
    Use `POST /api/v1/auth/refresh` to renew.

    ### Roles
    | Role | Description |
    |---|---|
    | `SEEKER` | Tenant searching for accommodation |
    | `LANDLORD` | Property owner managing listings |
    | `ADMIN` | Platform administrator |

    ### Base URL
    - Development: `http://localhost:8000`
    - Production: `https://api.nyumbalink.co.tz`
  contact:
    name: NyumbaLink Support
    email: mnemosadick@gmail.com
    url: https://nyumbalink.co.tz
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.nyumbalink.co.tz
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token. Obtain via `POST /api/v1/auth/login`.
        Include as: `Authorization: Bearer <token>`
  schemas:
...
    ApplicationCreate:
      type: object
      required: [property_id]
      properties:
        property_id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440001"
        move_in_date:
          type: string
          format: date
          nullable: true
          example: "2025-08-01"
        message:
          type: string
          nullable: true
          example: "I am a UDSM student looking for a quiet room."
    ApplicationOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seeker_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        application_status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, CANCELLED]
          example: "PENDING"
        move_in_date:
          type: string
          format: date
          nullable: true
        message:
          type: string
          nullable: true
        applied_at:
          type: string
          format: date-time
    ApplicationStatusUpdate:
      type: object
      required: [application_status]
      properties:
        application_status:
          type: string
          enum: [PENDING, APPROVED, REJECTED, CANCELLED]
          example: "APPROVED"
    AppointmentCreate:
      type: object
      required: [property_id, appointment_date]
      properties:
        property_id:
          type: string
          format: uuid
        appointment_date:
          type: string
          format: date-time
          example: "2025-07-20T10:00:00Z"
        notes:
          type: string
          nullable: true
          example: "Please confirm 24 hours before the visit."
    AppointmentOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seeker_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        appointment_date:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
          example: "PENDING"
        notes:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
    AppointmentStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [PENDING, CONFIRMED, CANCELLED, COMPLETED]
          example: "CONFIRMED"

    # ── Payments ──────────────────────────────────────────────────────────────
...
    ReviewCreate:
      type: object
      required: [property_id, rating]
      properties:
        property_id:
          type: string
          format: uuid
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 4
        review_text:
          type: string
          nullable: true
          example: "Very clean hostel, good security. Water supply is excellent."
    ReviewOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
        reviewer_id:
          type: string
          format: uuid
        property_id:
          type: string
          format: uuid
        landlord_id:
          type: string
          format: uuid
          nullable: true
        rating:
          type: integer
          nullable: true
          example: 4
        review_text:
          type: string
          nullable: true
        sentiment:
          type: string
          enum: [POSITIVE, NEUTRAL, NEGATIVE]
          nullable: true
          example: "POSITIVE"
        moderation_status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
          example: "PENDING"
        created_at:
          type: string
          format: date-time
    ReviewModerationUpdate:
      type: object
      required: [moderation_status]
      properties:
        moderation_status:
          type: string
          enum: [PENDING, APPROVED, REJECTED]
          example: "APPROVED"

    # ── Messaging ─────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/rentals/favorites/{property_id}:
    post:
      tags: [Rentals]
      summary: Save a property to favorites
      operationId: rentals_add_favorite
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "201":
          description: Added to favorites
        "409":
          description: Already in favorites
    delete:
      tags: [Rentals]
      summary: Remove a property from favorites
      operationId: rentals_remove_favorite
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "204":
          description: Removed from favorites
  /api/v1/rentals/applications:
    post:
      tags: [Rentals]
      summary: Submit a rental application
      description: Seeker applies for a property. Landlord must approve/reject.
      operationId: rentals_apply
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationCreate"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationOut"
...
  /api/v1/rentals/applications/property/{property_id}:
    get:
      tags: [Rentals]
      summary: List applications for a property (Landlord / Admin)
      operationId: rentals_property_applications
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "200":
          description: Applications for the property
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ApplicationOut"
  /api/v1/rentals/applications/{application_id}:
    patch:
      tags: [Rentals]
      summary: Update application status (Landlord / Admin)
      operationId: rentals_update_application
      parameters:
        - name: application_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationStatusUpdate"
      responses:
        "200":
          description: Updated application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationOut"
  /api/v1/rentals/appointments:
    post:
      tags: [Rentals]
      summary: Book a property viewing appointment
      operationId: rentals_book_appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentCreate"
      responses:
        "201":
          description: Appointment booked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentOut"
...
  /api/v1/rentals/appointments/{appointment_id}:
    patch:
      tags: [Rentals]
      summary: Update appointment status
      operationId: rentals_update_appointment
      parameters:
        - name: appointment_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AppointmentStatusUpdate"
      responses:
        "200":
          description: Updated appointment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentOut"

  # ────────────────────────────────── PAYMENTS ───────────────────────────────
...
  /api/v1/reviews:
    post:
      tags: [Reviews]
      summary: Submit a property review
      description: |
        Authenticated users can review properties they have visited or rented.
        Sentiment is auto-detected (rule-based; upgrade to OpenAI in production).
        All reviews start in `PENDING` moderation status.
      operationId: reviews_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewCreate"
      responses:
        "201":
          description: Review submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewOut"
  /api/v1/reviews/property/{property_id}:
    get:
      tags: [Reviews]
      summary: Get approved reviews for a property
      security: []
      operationId: reviews_get_property
      parameters:
        - $ref: "#/components/parameters/PropertyId"
      responses:
        "200":
          description: Approved reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewOut"
...
  /api/v1/reviews/{review_id}/moderate:
    patch:
      tags: [Reviews]
      summary: Approve or reject a review (Admin only)
      operationId: reviews_moderate
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewModerationUpdate"
      responses:
        "200":
          description: Moderated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewOut"

  # ────────────────────────────────── MESSAGING ──────────────────────────────
...